Add “include” User to StudyDefinition. Change the Study UI to show the name of the Principal Investigator
Add a route to take the study.  This returns the URL to go to for the frontend.
Change the way we fake the CHAOS session to make it a value we create and track here which is bound to the user information.  That way a specific frontend session can be tied to a user taking the study.
Fix a bug where two requests could be sent simultaneously when the access token was expired, locking up the DB when they both tried to refresh their respective tokens.

Clarify all the foreign key relationships in the DB & models.

Add config file for elicit config params.